Add check for rotate tag
When a video has a "rotate" tag in metadata, the video player rotates when playing the video, giving it a portrait perspective.
Because of this, the width is still larger than height even in a portrait video. Which causes the stride to be off and the resulting frame to be scrambled.
This can be fixed by inverting the widht and height of the video when there is a rotation.